Pizza Timer v1.2
Added Lap Flags
Added Custom Commands (windows only, commands must be written as if you're on the command line).
Changed the numbers to not be transparent and be more like in the game.
The Pizza Time timer should always appear at the bottom center of your screen. Same with the lap flags, but from the top center.